# contracts/openapi/content-service-v1.yaml
openapi: 3.1.0
info:
  title: Content Service Admin API
  version: 1.0.0
  description: Minimal Phase 1.5 admin integration contract. Development header authentication is non-production.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Deployment-relative; clients configure the environment-specific origin
paths:
...
  /api/v1/admin/exams:
    get: { operationId: listExams, summary: listExams, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], parameters: [{name: page, in: query, schema: {type: integer, default: 0}}, {name: size, in: query, schema: {type: integer, default: 20}}, {name: search, in: query, schema: {type: string}}, {name: status, in: query, schema: {$ref: '#/components/schemas/StructureStatus'}}, {name: countryCode, in: query, schema: {type: string}}, {name: sort, in: query, schema: {type: string}}], responses: {'200': {description: Exam page, content: {application/json: {schema: {$ref: '#/components/schemas/ExamPage'}}}}, '401': {$ref: '#/components/responses/AuthenticationRequired'}, '403': {$ref: '#/components/responses/Forbidden'}} }
    post: { operationId: createExam, summary: createExam, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ExamRequest'}}}}, responses: {'201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Exam'}}}}, '401': {$ref: '#/components/responses/AuthenticationRequired'}, '403': {$ref: '#/components/responses/Forbidden'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/Validation'}} }
  /api/v1/admin/exams/{examId}:
    parameters: [{$ref: '#/components/parameters/ExamId'}]
    get: {operationId: getExam, summary: getExam, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], responses: {'200': {description: Exam, content: {application/json: {schema: {$ref: '#/components/schemas/Exam'}}}}, '404': {$ref: '#/components/responses/NotFound'}}}
    put: {operationId: updateExam, summary: updateExam, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ExamRequest'}}}}, responses: {'200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/Exam'}}}}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/Validation'}}}
  /api/v1/admin/exams/{examId}/archive:
    parameters: [{$ref: '#/components/parameters/ExamId'}]
    post: {operationId: archiveExam, summary: archiveExam, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {$ref: '#/components/requestBodies/VersionOnly'}, responses: {'200': {description: Archived, content: {application/json: {schema: {$ref: '#/components/schemas/Exam'}}}}, '409': {$ref: '#/components/responses/Conflict'}}}
  /api/v1/admin/exams/{examId}/versions:
    parameters: [{$ref: '#/components/parameters/ExamId'}]
    get: {operationId: listExamVersions, summary: listExamVersions, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], parameters: [{name: status, in: query, schema: {$ref: '#/components/schemas/StructureStatus'}}, {name: validAt, in: query, schema: {type: string, format: date}}, {name: search, in: query, schema: {type: string}}], responses: {'200': {description: Versions, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/ExamVersion'}}}}}}}
    post: {operationId: createExamVersion, summary: createExamVersion, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ExamVersionRequest'}}}}, responses: {'201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/ExamVersion'}}}}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/Validation'}}}
  /api/v1/admin/exam-versions/{examVersionId}:
    parameters: [{$ref: '#/components/parameters/ExamVersionId'}]
    get: {operationId: getExamVersion, summary: getExamVersion, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], responses: {'200': {description: Version, content: {application/json: {schema: {$ref: '#/components/schemas/ExamVersion'}}}}, '404': {$ref: '#/components/responses/NotFound'}}}
    put: {operationId: updateExamVersion, summary: updateExamVersion, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ExamVersionRequest'}}}}, responses: {'200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/ExamVersion'}}}}, '409': {$ref: '#/components/responses/Conflict'}}}
  /api/v1/admin/exam-versions/{examVersionId}/archive:
    parameters: [{$ref: '#/components/parameters/ExamVersionId'}]
    post: {operationId: archiveExamVersion, summary: archiveExamVersion, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {$ref: '#/components/requestBodies/VersionOnly'}, responses: {'200': {description: Archived, content: {application/json: {schema: {$ref: '#/components/schemas/ExamVersion'}}}}}}
  /api/v1/admin/exam-versions/{examVersionId}/subjects:
    parameters: [{$ref: '#/components/parameters/ExamVersionId'}]
    get: {operationId: listSubjects, summary: listSubjects, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], responses: {'200': {description: Subjects, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Subject'}}}}}}}
    post: {operationId: createSubject, summary: createSubject, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/NodeRequest'}}}}, responses: {'201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Subject'}}}}}}
  /api/v1/admin/exam-versions/{examVersionId}/subjects/order:
    parameters: [{$ref: '#/components/parameters/ExamVersionId'}]
    put: {operationId: reorderSubjects, summary: reorderSubjects, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {$ref: '#/components/requestBodies/Order'}, responses: {'204': {description: Reordered}, '409': {$ref: '#/components/responses/Conflict'}}}
  /api/v1/admin/subjects/{subjectId}:
    parameters: [{$ref: '#/components/parameters/SubjectId'}]
    get: {operationId: getSubject, summary: getSubject, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], responses: {'200': {description: Subject, content: {application/json: {schema: {$ref: '#/components/schemas/Subject'}}}}}}
    put: {operationId: updateSubject, summary: updateSubject, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/NodeRequest'}}}}, responses: {'200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/Subject'}}}}, '409': {$ref: '#/components/responses/Conflict'}}}
  /api/v1/admin/subjects/{subjectId}/archive:
    parameters: [{$ref: '#/components/parameters/SubjectId'}]
    post: {operationId: archiveSubject, summary: archiveSubject, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {$ref: '#/components/requestBodies/VersionOnly'}, responses: {'200': {description: Archived, content: {application/json: {schema: {$ref: '#/components/schemas/Subject'}}}}}}
  /api/v1/admin/subjects/{subjectId}/topics:
    parameters: [{$ref: '#/components/parameters/SubjectId'}]
    get: {operationId: listTopics, summary: listTopics, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], responses: {'200': {description: Topics, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Topic'}}}}}}}
    post: {operationId: createTopic, summary: createTopic, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/NodeRequest'}}}}, responses: {'201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Topic'}}}}}}
  /api/v1/admin/subjects/{subjectId}/topics/order:
    parameters: [{$ref: '#/components/parameters/SubjectId'}]
    put: {operationId: reorderTopics, summary: reorderTopics, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {$ref: '#/components/requestBodies/Order'}, responses: {'204': {description: Reordered}, '409': {$ref: '#/components/responses/Conflict'}}}
  /api/v1/admin/topics/{topicId}:
    parameters: [{$ref: '#/components/parameters/TopicId'}]
    get: {operationId: getTopic, summary: getTopic, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], responses: {'200': {description: Topic, content: {application/json: {schema: {$ref: '#/components/schemas/Topic'}}}}}}
    put: {operationId: updateTopic, summary: updateTopic, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/NodeRequest'}}}}, responses: {'200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/Topic'}}}}, '409': {$ref: '#/components/responses/Conflict'}}}
  /api/v1/admin/topics/{topicId}/archive:
    parameters: [{$ref: '#/components/parameters/TopicId'}]
    post: {operationId: archiveTopic, summary: archiveTopic, security: [{developmentAdminIdentity: [], developmentAdminRoles: []}], requestBody: {$ref: '#/components/requestBodies/VersionOnly'}, responses: {'200': {description: Archived, content: {application/json: {schema: {$ref: '#/components/schemas/Topic'}}}}}}
...
components:
  securitySchemes:
    developmentAdminIdentity:
      type: apiKey
      in: header
      name: X-Admin-Identity
      description: Non-production identity header.
    developmentAdminRoles:
      type: apiKey
      in: header
      name: X-Admin-Roles
      description: Non-production comma-separated roles. Recognized roles are CONTENT_AUTHOR, CONTENT_REVIEWER, CONTENT_PUBLISHER, and ADMIN.
  parameters:
    ExamId: {name: examId, in: path, required: true, schema: {type: string, format: uuid}}
    ExamVersionId: {name: examVersionId, in: path, required: true, schema: {type: string, format: uuid}}
    SubjectId: {name: subjectId, in: path, required: true, schema: {type: string, format: uuid}}
    TopicId: {name: topicId, in: path, required: true, schema: {type: string, format: uuid}}
...
  requestBodies:
    VersionOnly: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/VersionOnly'}}}}
    Order: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/OrderRequest'}}}}
...
  responses:
    AuthenticationRequired:
      description: Valid admin authentication is required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: The authenticated identity lacks a required role
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    Validation: {description: Request validation failed,content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    NotFound: {description: Resource not found,content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
    Conflict: {description: Conflict or stale version,content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}}
  schemas:
    StructureStatus: {type: string, enum: [DRAFT, ACTIVE, ARCHIVED]}
...
    VersionOnly: {type: object,required: [version],properties: {version: {type: integer,format: int64,minimum: 0}}}
    OrderRequest: {type: object,required: [ids],properties: {ids: {type: array,items: {type: string,format: uuid}}}}
    AuditFields: {type: object,required: [id,createdAt,updatedAt,version],properties: {id: {type: string,format: uuid},createdAt: {type: string,format: date-time},updatedAt: {type: string,format: date-time},version: {type: integer,format: int64}}}
    Exam: {allOf: [{$ref: '#/components/schemas/AuditFields'},{type: object,required: [code,name,countryCode,status,versionCount],properties: {code: {type: string},name: {type: string},countryCode: {type: string},status: {$ref: '#/components/schemas/StructureStatus'},versionCount: {type: integer,format: int64}}}]}
    ExamRequest: {type: object,required: [code,name,countryCode,status],properties: {code: {type: string},name: {type: string},countryCode: {type: string},status: {$ref: '#/components/schemas/StructureStatus'},version: {type: integer,format: int64}}}
    ExamPage: {type: object,required: [items,page,size,totalItems,totalPages],properties: {items: {type: array,items: {$ref: '#/components/schemas/Exam'}},page: {type: integer},size: {type: integer},totalItems: {type: integer,format: int64},totalPages: {type: integer,format: int64}}}
    ExamVersion: {allOf: [{$ref: '#/components/schemas/AuditFields'},{type: object,required: [examId,versionCode,displayName,status],properties: {examId: {type: string,format: uuid},versionCode: {type: string},displayName: {type: string},status: {$ref: '#/components/schemas/StructureStatus'},validFrom: {type: [string,'null'],format: date},validTo: {type: [string,'null'],format: date}}}]}
    ExamVersionRequest: {type: object,required: [versionCode,displayName,status],properties: {versionCode: {type: string},displayName: {type: string},status: {$ref: '#/components/schemas/StructureStatus'},validFrom: {type: [string,'null'],format: date},validTo: {type: [string,'null'],format: date},version: {type: integer,format: int64}}}
    NodeRequest: {type: object,required: [code,name,sortOrder,status],properties: {code: {type: string},name: {type: string},description: {type: [string,'null']},sortOrder: {type: integer,minimum: 0},status: {$ref: '#/components/schemas/StructureStatus'},version: {type: integer,format: int64}}}
    Subject: {allOf: [{$ref: '#/components/schemas/AuditFields'},{$ref: '#/components/schemas/NodeRequest'},{type: object,required: [examVersionId],properties: {examVersionId: {type: string,format: uuid}}}]}
    Topic: {allOf: [{$ref: '#/components/schemas/AuditFields'},{$ref: '#/components/schemas/NodeRequest'},{type: object,required: [subjectId],properties: {subjectId: {type: string,format: uuid}}}]}
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, timestamp, errors]
      properties:
        code: { type: string }
        message: { type: string }
        timestamp: { type: string, format: date-time }
        errors:
          type: array
          items:
            type: object
            additionalProperties: true